+++ src/Sulu/Bundle/ResourceBundle/Resources/public/js/components/filters/util/toolbar-extension-handler.js
@@ -14,7 +14,7 @@ define(['app-config', 'config'], function(AppConfig, Config) {
     var constants = {
             filterListUrl: 'resource/filters/',
             manageFilters: 'manage',
-            filtersUrl: 'api/filters?flat=true&context=',
+            filtersUrl: 'api/filters?flat=true&limit=100&context=',
             filterUrl: 'resource/filters/',
             toolbarSelectButtonId: 'filters'
         },
@@ -32,6 +32,7 @@ define(['app-config', 'config'], function(AppConfig, Config) {
             if (!!context && !!toolbarItems && !!dataGridInstanceName && !!this.config.contexts[context]) {
                 var url = constants.filtersUrl + context,
                     updateEventName = 'husky.datagrid.' + dataGridInstanceName + '.updated',
+                    errorEventName = 'husky.datagrid.' + dataGridInstanceName + '.loading.failed',
                     filterDropDown = getFilterDropdown.call(this, context, dataGridInstanceName, url);
 
                 this.filterResultSelector = filterResultSelector;
@@ -42,6 +43,8 @@ define(['app-config', 'config'], function(AppConfig, Config) {
                 toolbarItems.push(filterDropDown);
                 this.sandbox.off(updateEventName);
                 this.sandbox.on(updateEventName, updateFilterResult.bind(this));
+
+                this.sandbox.on(errorEventName, handleLoadingError.bind(this));
             }
         },
 
@@ -82,6 +85,18 @@ define(['app-config', 'config'], function(AppConfig, Config) {
             } else {
                 this.sandbox.logger.log('Either no filter is set or the result contains no total number!');
             }
+        },
+
+        handleLoadingError = function() {
+            this.sandbox.emit('sulu.labels.error.show',
+                this.sandbox.translate('resource.filter.error-loading'),
+                'labels.error',
+                ''
+            );
+
+            this.sandbox.emit('husky.datagrid.' + this.datagridInstance + '.url.update', {filter: ''});
+
+            unsetFilter.call(this);
         },
 
         /**